Moved parametrized tests in a separate crate
