+++ episode/6/balanced/Cargo.toml
@@ -4,6 +4,3 @@ version = "0.1.0"
 authors = ["Matthias Endler <matthias-endler@gmx.net>"]
 
 [dependencies]
-
-[dev-dependencies]
-rstest="*"
